README.md: notify that a device may require a set of independent modules
Who are not familiar in the DVB subsystem might not be aware of that fact
that a real device may require a set of modules to operate which might
are not selected together automatically when one part selected.
A short paragraph in the README can save those users from confusion.