README.md: clarify Kernel modules > Video Support conflict
Using media modules from the mainline kernel is not recommended when
v4l-dvb is in use, because it is very likely that conflict will arise.

If it becomes necessary it'd be still better to provide camera support
based on used media tree rather going into a complex interaction with
mainline media modules.