oscam: fix generated tarball hash
Previously the generated tarball differed beacuse checkout date was used
to create directories inside archive. The recent changes in OpenWrt
force date of last SVN commit as timestamp for dirs inside archive, so
finaly the checksum won't change on each rebuild of the archive. Update
check sum to reflect that.

Signed-off-by: Tomasz Maciej Nowak <tmn505@gmail.com>